Shifter FBX tool: add anim clips module
