+++ release/scripts/mgear/shifter/game_tools_fbx/partition_widgets.py
@@ -1,16 +1,11 @@
 import copy
 import weakref
-from functools import partial
 
-from mgear.vendor.Qt import QtWidgets
-from mgear.vendor.Qt import QtCore
-from mgear.vendor.Qt import QtGui
+from mgear.vendor.Qt import QtWidgets, QtCore, QtGui
 
 from mgear.core import pyqt, utils
-from mgear.shifter.game_tools_fbx import (
-    fbx_export_node,
-    utils as fbx_utils
-)
+from mgear.shifter.game_tools_fbx import fbx_export_node
+
 
 import maya.cmds as cmds
 
@@ -286,7 +281,6 @@ class TreeItem(QtWidgets.QTreeWidgetItem):
 
 
 class OutlinerTreeView(QtWidgets.QTreeWidget):
-
     TREE_ITEM_CLASS = TreeItem
     NODE_CLASS = NodeClass
 
@@ -450,7 +444,6 @@ class OutlinerTreeView(QtWidgets.QTreeWidget):
         self.reset_contents()
 
     def can_be_dropped(self, index):
-
         item = self.itemFromIndex(index)
         drop_indicator_position = self.dropIndicatorPosition()
 
@@ -480,7 +473,6 @@ class OutlinerTreeView(QtWidgets.QTreeWidget):
         return dict()
 
     def populate_items(self, add_callbacks=True):
-
         if add_callbacks:
             self.cleanup()
 
@@ -539,7 +531,6 @@ class OutlinerTreeView(QtWidgets.QTreeWidget):
         self._registered_node_callbacks = list()
 
     def _create_root_item(self, partition):
-
         node_icon = pyqt.get_icon("mgear_package")
         root_node = self.NODE_CLASS(
             partition, "Root", True, node_icon, True, True
@@ -550,7 +541,6 @@ class OutlinerTreeView(QtWidgets.QTreeWidget):
         return item
 
     def _add_partition_item(self, node, partition_item):
-
         node_icon = pyqt.get_icon("mgear_box")
         item_node = self.NODE_CLASS(
             node,
@@ -743,7 +733,6 @@ class TreeViewDelegate(QtWidgets.QItemDelegate):
 
 
 class RowPainter(object):
-
     DISABLED_BACKGROUND_IMAGE = fx.getPixmap("out_MASH_ChevronBG")
     DISABLED_HIGHLIGHT_IMAGE = fx.getPixmap("out_MASH_ChevronBGSelected")
     EXPANDED_ARROW = (
@@ -1022,529 +1011,3 @@ class RowPainter(object):
             self._draw_action(
                 action_name, pixmap, self._rect.right() - start, top
             )
-
-
-class AnimClipsListWidget(QtWidgets.QWidget):
-    def __init__(self, parent=None):
-        super(AnimClipsListWidget, self).__init__(parent=parent)
-
-        self._game_tools_fbx = parent
-        self._root_joint = None
-        self._export_path = None
-        self._file_name = None
-
-        main_layout = QtWidgets.QVBoxLayout()
-        self.setLayout(main_layout)
-
-        anim_clip_option_layout = QtWidgets.QHBoxLayout()
-        self._anim_clips_checkbox = QtWidgets.QCheckBox(parent=self)
-        self._anim_clips_checkbox.setChecked(True)
-        self._add_anim_clip_button = QtWidgets.QPushButton(
-            "Add Clip", parent=self
-        )
-        self._delete_all_clips_button = QtWidgets.QPushButton(
-            "Delete All Clips", parent=self
-        )
-        anim_clip_option_layout.addWidget(self._anim_clips_checkbox)
-        anim_clip_option_layout.addStretch()
-        anim_clip_option_layout.addWidget(self._add_anim_clip_button)
-        anim_clip_option_layout.addWidget(self._delete_all_clips_button)
-
-        anim_clip_area = QtWidgets.QScrollArea(parent=self)
-        anim_clip_area.setWidgetResizable(True)
-        anim_clip_content = QtWidgets.QWidget(parent=self)
-        anim_clip_area.setWidget(anim_clip_content)
-
-        anim_clip_stretch_layout = QtWidgets.QVBoxLayout()
-        anim_clip_stretch_layout.setSpacing(0)
-        anim_clip_stretch_layout.setContentsMargins(0, 0, 0, 0)
-        self._clips_layout = QtWidgets.QVBoxLayout()
-        self._clips_layout.setSpacing(0)
-        self._clips_layout.setContentsMargins(0, 0, 0, 0)
-        anim_clip_stretch_layout.addLayout(self._clips_layout)
-        anim_clip_content.setLayout(anim_clip_stretch_layout)
-        anim_clip_stretch_layout.addStretch()
-
-        main_layout.addLayout(anim_clip_option_layout)
-        main_layout.addWidget(anim_clip_area)
-
-        self._anim_clips_checkbox.toggled.connect(
-            self._on_toggled_anim_clips_checkbox
-        )
-        self._add_anim_clip_button.clicked.connect(
-            self._on_add_animation_clip_button_clicked
-        )
-        self._delete_all_clips_button.clicked.connect(
-            self._on_delete_all_animation_clips_button_clicked
-        )
-
-    def refresh(self):
-
-        pyqt.clear_layout(self._clips_layout)
-
-        if not self._game_tools_fbx:
-            return
-        self._root_joint = self._game_tools_fbx.get_root_joint()
-        self._export_path = self._game_tools_fbx.get_export_path()
-        self._file_name = self._game_tools_fbx.get_file_name()
-
-        if not self._root_joint:
-            return
-        export_node = fbx_export_node.FbxExportNode.get()
-        if not export_node:
-            return
-
-        for anim_clip_data in export_node.get_animation_clips(
-            self._root_joint
-        ):
-            anim_clip_name = anim_clip_data.get("title", None)
-            if not anim_clip_name:
-                continue
-            self._add_animation_clip(anim_clip_name)
-
-    def _add_animation_clip(self, clip_name):
-
-        # TODO: Add code to avoid adding anim clips with duplicated names
-
-        anim_clip_widget = AnimClipWidget(
-            clip_name, parent=self._game_tools_fbx
-        )
-        self._clips_layout.addWidget(anim_clip_widget)
-
-        return anim_clip_widget
-
-    def _on_toggled_anim_clips_checkbox(self, flag):
-
-        for i in range(self._clips_layout.count()):
-            anim_clip_widget = self._clips_layout.itemAt(i).widget()
-            anim_clip_widget.set_enabled(flag)
-
-    def _on_add_animation_clip_button_clicked(self):
-
-        if not self._root_joint:
-            cmds.warning(
-                "Was not possible to add animation clip because not root joint is defined!"
-            )
-            return
-
-        export_node = fbx_export_node.FbxExportNode.get() or fbx_export_node.FbxExportNode.create()
-        anim_clip_name = export_node.add_animation_clip(self._root_joint)
-        if not anim_clip_name:
-            cmds.warning("Was not possible to add new animation clip")
-            return
-
-        self._add_animation_clip(anim_clip_name)
-
-    def _on_delete_all_animation_clips_button_clicked(self):
-
-        if not self._root_joint:
-            return
-
-        export_node = fbx_export_node.FbxExportNode.get()
-        if not export_node:
-            return
-
-        export_node.delete_all_animation_clips()
-
-        self.refresh()
-
-
-class AnimClipWidget(QtWidgets.QFrame):
-    def __init__(self, clip_name=None, parent=None):
-        super(AnimClipWidget, self).__init__(parent)
-
-        self._game_tools_fbx = parent
-        self._clip_name = clip_name
-        self._previous_name = clip_name
-
-        self.setFixedHeight(150)
-        self.window().setAttribute(QtCore.Qt.WA_AlwaysShowToolTips, True)
-        self.setContextMenuPolicy(QtCore.Qt.CustomContextMenu)
-
-        main_layout = QtWidgets.QVBoxLayout()
-        self.setLayout(main_layout)
-
-        self._export_checkbox = QtWidgets.QCheckBox(parent=self)
-        self._title_line_edit = QtWidgets.QLineEdit(parent=self)
-        self._title_line_edit.setSizePolicy(
-            QtWidgets.QSizePolicy.MinimumExpanding,
-            QtWidgets.QSizePolicy.Preferred,
-        )
-        self._close_button = QtWidgets.QPushButton("X", parent=self)
-        self._close_button.setMaximumWidth(20)
-        close_layout = QtWidgets.QHBoxLayout()
-        close_layout.setSpacing(2)
-        close_layout.setContentsMargins(0, 0, 0, 0)
-        close_layout.addWidget(self._export_checkbox)
-        close_layout.addWidget(self._title_line_edit)
-        close_layout.addWidget(self._close_button)
-
-        options_layout_1 = QtWidgets.QHBoxLayout()
-        options_layout_1.setSpacing(2)
-        options_layout_1.setContentsMargins(0, 0, 0, 0)
-        # self._start_at_origin_checkbox = QtWidgets.QCheckBox('Origin', parent=self)
-        # self._start_at_frame_zero_checkbox = QtWidgets.QCheckBox('Frame 0', parent=self)
-        # options_layout_1.addWidget(self._export_checkbox)
-        # options_layout_1.addWidget(self._start_at_origin_checkbox)
-        # options_layout_1.addWidget(self._start_at_frame_zero_checkbox)
-        self._export_checkbox.setChecked(True)
-        # self._start_at_origin_checkbox.setChecked(True)
-
-        options_layout_2 = QtWidgets.QHBoxLayout()
-        options_layout_2.setSpacing(2)
-        options_layout_2.setContentsMargins(0, 0, 0, 0)
-        self._up_button = QtWidgets.QPushButton(parent=self)
-        up_pix = QtGui.QPixmap(":UVTkArrowRight.png")
-        rm_up_pix = QtGui.QMatrix()
-        rm_up_pix.rotate(-90)
-        up_pix = up_pix.transformed(rm_up_pix)
-        self._up_button.setIcon(up_pix)
-        self._up_button.setStatusTip("Move Sequence Up")
-        self._down_button = QtWidgets.QPushButton(parent=self)
-        down_pix = QtGui.QPixmap(":UVTkArrowRight.png")
-        rm_down_pix = QtGui.QMatrix()
-        rm_down_pix.rotate(90)
-        down_pix = down_pix.transformed(rm_down_pix)
-        self._down_button.setIcon(down_pix)
-        self._down_button.setStatusTip("Move Sequence Down")
-        self._set_range_button = QtWidgets.QPushButton(parent=self)
-        self._set_range_button.setIcon(QtGui.QIcon(":adjustTimeline.png"))
-        self._set_range_button.setStatusTip("Set Range")
-        self._playbast_button = QtWidgets.QPushButton(parent=self)
-        self._playbast_button.setIcon(QtGui.QIcon(":playblast.png"))
-        self._playbast_button.setStatusTip("Playblast")
-        self._play_button = QtWidgets.QPushButton(parent=self)
-        self._play_button.setIcon(QtGui.QIcon(":playClip.png"))
-        self._play_button.setStatusTip("Play Sequence")
-        self._export_button = QtWidgets.QPushButton("Export", parent=self)
-        options_layout_2.addWidget(self._up_button)
-        options_layout_2.addWidget(self._down_button)
-        options_layout_2.addWidget(self._set_range_button)
-        options_layout_2.addWidget(self._playbast_button)
-        options_layout_2.addWidget(self._play_button)
-        options_layout_2.addWidget(self._export_button)
-
-        options_layout_3 = QtWidgets.QHBoxLayout()
-        options_layout_3.setSpacing(2)
-        options_layout_3.setContentsMargins(0, 0, 0, 0)
-        start_frame_label = QtWidgets.QLabel("Start Frame:", parent=self)
-        self._start_frame_line_edit = QtWidgets.QLineEdit(parent=self)
-        self._start_frame_line_edit.setValidator(QtGui.QIntValidator())
-        end_frame_label = QtWidgets.QLabel("End Frame:", parent=self)
-        self._end_frame_line_edit = QtWidgets.QLineEdit(parent=self)
-        self._end_frame_line_edit.setValidator(QtGui.QIntValidator())
-        self._frame_rate_combo = QtWidgets.QComboBox(parent=self)
-        self._frame_rate_combo.addItems(["30 FPS", "60 FPS", "120 FPS"])
-        options_layout_3.addWidget(start_frame_label)
-        options_layout_3.addWidget(self._start_frame_line_edit)
-        options_layout_3.addWidget(end_frame_label)
-        options_layout_3.addWidget(self._end_frame_line_edit)
-        options_layout_3.addWidget(self._frame_rate_combo)
-
-        options_layout_4 = QtWidgets.QHBoxLayout()
-        options_layout_4.setSpacing(2)
-        options_layout_4.setContentsMargins(0, 0, 0, 0)
-        self._anim_layer_label = QtWidgets.QLabel(
-            "Animation Layer: ", parent=self
-        )
-        self._anim_layer_combo = QtWidgets.QComboBox(parent=self)
-        self._anim_layer_combo.setSizePolicy(
-            QtWidgets.QSizePolicy.MinimumExpanding,
-            QtWidgets.QSizePolicy.Maximum,
-        )
-        options_layout_4.addWidget(self._anim_layer_label)
-        options_layout_4.addWidget(self._anim_layer_combo)
-
-        main_layout.addLayout(close_layout)
-        main_layout.addLayout(options_layout_1)
-        main_layout.addLayout(options_layout_2)
-        main_layout.addLayout(options_layout_3)
-        main_layout.addLayout(options_layout_4)
-
-        self.refresh()
-
-        self._close_button.clicked.connect(self._on_close_button_clicked)
-        self._title_line_edit.textChanged.connect(self._on_update_anim_clip)
-        self._export_checkbox.toggled.connect(self._on_update_anim_clip)
-        # self._start_at_origin_checkbox.toggled.connect(self._on_update_anim_clip)
-        # self._start_at_frame_zero_checkbox.toggled.connect(self._on_update_anim_clip)
-        self._start_frame_line_edit.textChanged.connect(
-            self._on_update_anim_clip
-        )
-        self._end_frame_line_edit.textChanged.connect(
-            self._on_update_anim_clip
-        )
-        self._frame_rate_combo.currentIndexChanged.connect(
-            self._on_update_anim_clip
-        )
-        self._anim_layer_combo.currentIndexChanged.connect(
-            self._on_update_anim_clip
-        )
-        self._set_range_button.clicked.connect(
-            self._on_set_range_button_clicked
-        )
-        self._playbast_button.clicked.connect(self._on_playbast_button_clicked)
-        self._play_button.clicked.connect(self._on_play_button_clicked)
-        self._export_button.clicked.connect(self._on_export_button_clicked)
-        self.customContextMenuRequested.connect(
-            self._on_custom_context_menu_requested
-        )
-
-    def refresh(self):
-        export_node = fbx_export_node.FbxExportNode.get()
-        if not export_node:
-            self._clear()
-            return
-
-        if not self._game_tools_fbx:
-            self._clear()
-            return
-
-        root_joint = self._game_tools_fbx.get_root_joint()
-        anim_clip_data = export_node.get_animation_clip(
-            root_joint, self._clip_name
-        )
-
-        with pyqt.block_signals(self._title_line_edit):
-            self._title_line_edit.setText(
-                anim_clip_data.get("title", "Untitled")
-            )
-        with pyqt.block_signals(self._start_frame_line_edit):
-            self._start_frame_line_edit.setText(
-                str(anim_clip_data.get("startFrame", ""))
-            )
-        with pyqt.block_signals(self._end_frame_line_edit):
-            self._end_frame_line_edit.setText(
-                str(anim_clip_data.get("endFrame", ""))
-            )
-        with pyqt.block_signals(self._frame_rate_combo):
-            self._frame_rate_combo.setCurrentIndex(
-                self._frame_rate_combo.findText(
-                    anim_clip_data.get("frameRate", "")
-                )
-            )
-        with pyqt.block_signals(self._export_checkbox):
-            self._export_checkbox.setChecked(
-                anim_clip_data.get("enabled", True)
-            )
-        # with pyqt.block_signals(self._start_at_origin_checkbox):
-        # 	self._start_at_origin_checkbox.setChecked(anim_clip_data.get('startOrigin', True))
-        # with pyqt.block_signals(self._start_at_frame_zero_checkbox):
-        # 	self._start_at_frame_zero_checkbox.setChecked(anim_clip_data.get('frameZero', False))
-
-        with pyqt.block_signals(self._anim_layer_combo):
-            self._anim_layer_combo.clear()
-            # TODO: Maybe we should filter display layers that are set with override mode?
-            anim_layers = fbx_utils.all_anim_layers_ordered(
-                include_base_animation=False
-            )
-            self._anim_layer_combo.addItems(["None"] + anim_layers)
-            self._anim_layer_combo.setCurrentText(
-                anim_clip_data.get("animLayer", "None")
-            )
-
-    def set_enabled(self, flag):
-        self._export_checkbox.setChecked(flag)
-
-    def toggle_enabled(self):
-        self._export_checkbox.setChecked(not self._export_checkbox.isChecked())
-
-    def _clear(self):
-        self._title_line_edit.setText("Untitled")
-        self._start_frame_line_edit.setText("")
-        self._end_frame_line_edit.setText("")
-        self._frame_rate_combo.setCurrentIndex(0)
-        self._export_checkbox.setChecked(False)
-        # self._start_at_origin_checkbox.setChecked(False)
-        # self._start_at_frame_zero_checkbox.setChecked(False)
-
-    def _on_close_button_clicked(self):
-        export_node = fbx_export_node.FbxExportNode.get()
-        if not export_node:
-            return
-
-        root_joint = (
-            self._game_tools_fbx.get_root_joint()
-            if self._game_tools_fbx
-            else None
-        )
-        if not root_joint:
-            return
-
-        export_node.delete_animation_clip(root_joint, self._clip_name)
-
-        self.setParent(None)
-        self.deleteLater()
-
-    def _on_update_anim_clip(self):
-
-        root_joint = (
-            self._game_tools_fbx.get_root_joint()
-            if self._game_tools_fbx
-            else None
-        )
-        if not root_joint:
-            return
-
-        export_node = fbx_export_node.FbxExportNode.get()
-        if not export_node:
-            return
-
-        anim_clip_data = fbx_export_node.FbxExportNode.ANIM_CLIP_DATA.copy()
-        anim_clip_data["title"] = self._title_line_edit.text()
-        # anim_clip_data['path'] = self._export_path_line_edit.text()
-        anim_clip_data["enabled"] = self._export_checkbox.isChecked()
-        # anim_clip_data['startOrigin'] = self._start_at_origin_checkbox.isChecked()
-        # anim_clip_data['frameZero'] = self._start_at_frame_zero_checkbox.isChecked()
-        anim_clip_data["frameRate"] = self._frame_rate_combo.currentText()
-        anim_clip_data["startFrame"] = int(self._start_frame_line_edit.text())
-        anim_clip_data["endFrame"] = int(self._end_frame_line_edit.text())
-        anim_layer = self._anim_layer_combo.currentText()
-        anim_clip_data["animLayer"] = (
-            anim_layer if anim_layer and anim_layer != "None" else ""
-        )
-
-        export_node.update_animation_clip(
-            root_joint, self._previous_name, anim_clip_data
-        )
-        self._previous_name = anim_clip_data["title"]
-
-    def _on_delete_anim_clip(self):
-        export_node = fbx_export_node.FbxExportNode.get()
-        if not export_node:
-            return
-
-        result = export_node.delete_animation_clip(
-            self._root_joint, self._clip_name
-        )
-        if not result:
-            return
-
-        self.setParent(None)
-        self.deleteLater()
-
-    def _on_set_range_button_clicked(self):
-
-        start_frame, end_frame = int(self._start_frame_line_edit.text()), int(
-            self._end_frame_line_edit.text()
-        )
-        cmds.playbackOptions(
-            animationStartTime=start_frame,
-            minTime=start_frame,
-            animationEndTime=end_frame,
-            maxTime=end_frame,
-        )
-        cmds.currentTime(start_frame, edit=True)
-
-    def _on_playbast_button_clicked(self):
-
-        start_frame, end_frame = int(self._start_frame_line_edit.text()), int(
-            self._end_frame_line_edit.text()
-        )
-        fbx_utils.create_mgear_playblast(
-            folder=self._export_path_line_edit.text(),
-            start_frame=start_frame,
-            end_frame=end_frame,
-        )
-
-    def _on_play_button_clicked(self):
-        start_frame, end_frame = int(self._start_frame_line_edit.text()), int(
-            self._end_frame_line_edit.text()
-        )
-        cmds.playbackOptions(
-            animationStartTime=start_frame,
-            minTime=start_frame,
-            animationEndTime=end_frame,
-            maxTime=end_frame,
-        )
-        if cmds.play(query=True, state=True):
-            cmds.play(state=False)
-        else:
-            cmds.play(forward=True)
-
-    def _on_export_button_clicked(self):
-
-        root_joint = (
-            self._game_tools_fbx.get_root_joint()
-            if self._game_tools_fbx
-            else None
-        )
-        if not root_joint:
-            cmds.error("No valid root joint defined!")
-            return False
-
-        export_path = (
-            self._game_tools_fbx.get_export_path()
-            if self._game_tools_fbx
-            else ""
-        )
-        file_name = (
-            self._game_tools_fbx.get_file_name()
-            if self._game_tools_fbx
-            else ""
-        )
-        remove_namespace = (
-            self._game_tools_fbx.get_remove_namespace()
-            if self._game_tools_fbx
-            else False
-        )
-        scene_clean = (
-            self._game_tools_fbx.get_scene_clean()
-            if self._game_tools_fbx
-            else False
-        )
-        anim_layer = self._anim_layer_combo.currentText()
-        anim_layer = anim_layer if anim_layer and anim_layer != "None" else ""
-
-        export_kwargs = {
-            "startFrame": int(self._start_frame_line_edit.text()),
-            "endFrame": int(self._end_frame_line_edit.text()),
-            "enabled": self._export_checkbox.isChecked(),
-            # 'startOrigin': self._start_at_origin_checkbox.isChecked(),
-            # 'frameZero': self._start_at_frame_zero_checkbox.isChecked(),
-            "frameRate": self._frame_rate_combo.currentText(),
-            "file_path": export_path,
-            "file_name": "{}_{}".format(
-                file_name, self._title_line_edit.text().replace(" ", "_")
-            ),
-            "remove_namespace": remove_namespace,
-            "scene_clean": scene_clean,
-            "anim_layer": anim_layer,
-        }
-        return fbx_utils.export_animation_clip(root_joint, **export_kwargs)
-
-    def _on_custom_context_menu_requested(self, pos):
-
-        context_menu = QtWidgets.QMenu(parent=self)
-        delete_anim_clip_action = context_menu.addAction("Delete")
-        playblast_menu = QtWidgets.QMenu("Playblast", parent=self)
-        playblast_25_action = playblast_menu.addAction("25%")
-        playblast_50_action = playblast_menu.addAction("50%")
-        playblast_75_action = playblast_menu.addAction("75%")
-        playblast_100_action = playblast_menu.addAction("100%")
-        playblast_menu.addSeparator()
-        open_playblasts_folder_action = playblast_menu.addAction(
-            "Open in Explorer"
-        )
-
-        context_menu.addAction(delete_anim_clip_action)
-        context_menu.addMenu(playblast_menu)
-
-        delete_anim_clip_action.triggered.connect(self._on_delete_anim_clip)
-        playblast_25_action.triggered.connect(
-            partial(fbx_utils.create_mgear_playblast, scale=25)
-        )
-        playblast_50_action.triggered.connect(
-            partial(fbx_utils.create_mgear_playblast, scale=50)
-        )
-        playblast_75_action.triggered.connect(
-            partial(fbx_utils.create_mgear_playblast, scale=75)
-        )
-        playblast_100_action.triggered.connect(
-            partial(fbx_utils.create_mgear_playblast, scale=100)
-        )
-        open_playblasts_folder_action.triggered.connect(
-            fbx_utils.open_mgear_playblast_folder
-        )
-
-        context_menu.exec_(self.mapToGlobal(pos))
