+++ release/scripts/mgear/shifter/game_tools_fbx/partitions_outliner.py
@@ -3,15 +3,20 @@ import maya.cmds as cmds
 from mgear.vendor.Qt import QtCore
 
 from mgear.core import pyqt
-from mgear.shifter.game_tools_fbx import (
-    fbx_export_node,
-    widgets
-)
+from mgear.shifter.game_tools_fbx import fbx_export_node, partition_widgets
 
 
-class PartitionNodeClass(widgets.NodeClass):
-    def __init__(self, node_name, node_type, is_root, icon, enabled,
-        network_enabled, is_master=False):
+class PartitionNodeClass(partition_widgets.NodeClass):
+    def __init__(
+        self,
+        node_name,
+        node_type,
+        is_root,
+        icon,
+        enabled,
+        network_enabled,
+        is_master=False,
+    ):
         super(PartitionNodeClass, self).__init__(
             node_name=node_name,
             node_type=node_type,
@@ -30,7 +35,8 @@ class PartitionNodeClass(widgets.NodeClass):
     def is_master(self, flag):
         self._is_master = flag
 
-class PartitionTreeItem(widgets.TreeItem):
+
+class PartitionTreeItem(partition_widgets.TreeItem):
     def __init__(self, node, header, show_enabled, parent=None):
         super(PartitionTreeItem, self).__init__(
             node=node, header=header, show_enabled=show_enabled, parent=parent
@@ -39,8 +45,8 @@ class PartitionTreeItem(widgets.TreeItem):
     def is_master(self):
         return self.node.is_master
 
-class PartitionsOutliner(widgets.OutlinerTreeView):
 
+class PartitionsOutliner(partition_widgets.OutlinerTreeView):
     NODE_CLASS = PartitionNodeClass
     TREE_ITEM_CLASS = PartitionTreeItem
 
@@ -58,7 +64,6 @@ class PartitionsOutliner(widgets.OutlinerTreeView):
         self.droppedItems.connect(self.partition_items_dropped)
 
     def mouseDoubleClickEvent(self, event):
-
         if event.button() == QtCore.Qt.LeftButton:
             index = self.indexAt(event.pos())
             if index.row() == -1:
@@ -74,7 +79,6 @@ class PartitionsOutliner(widgets.OutlinerTreeView):
         super(PartitionsOutliner, self).mouseDoubleClickEvent(event)
 
     def can_be_dropped(self, index):
-
         valid = super(PartitionsOutliner, self).can_be_dropped(index)
         if not valid:
             return valid
@@ -105,16 +109,17 @@ class PartitionsOutliner(widgets.OutlinerTreeView):
         export_nodes = fbx_export_node.FbxExportNode.find()
         if not export_nodes:
             return dict()
+        export_node = export_nodes[0]
         if len(export_nodes) > 2:
             cmds.warning(
-                'Multiple FBX Export nodes found in scene. Using first one found: "{}"'.format(
-                    export_nodes[0]
+                'Multiple FBX Export nodes found in scene. \
+                         Using first one found: "{}"'.format(
+                    export_node
                 )
             )
-        return export_nodes[0].get_partitions()
+        return export_node.get_partitions()
 
     def populate_items(self, add_callbacks=True):
-
         if add_callbacks:
             self.cleanup()
 
@@ -160,7 +165,6 @@ class PartitionsOutliner(widgets.OutlinerTreeView):
         self._update_master_partition()
 
     def _on_custom_context_menu_requested(self, pos):
-
         item = self._get_current_item()
         if not item or item.is_master():
             return
@@ -168,7 +172,6 @@ class PartitionsOutliner(widgets.OutlinerTreeView):
         super(PartitionsOutliner, self)._on_custom_context_menu_requested(pos)
 
     def _create_master_root_item(self):
-
         node_icon = pyqt.get_icon("mgear_package")
         root_node = self.NODE_CLASS(
             "Master", "Root", True, node_icon, True, True, is_master=True
@@ -183,7 +186,6 @@ class PartitionsOutliner(widgets.OutlinerTreeView):
         return item
 
     def _add_master_partition_item(self, node, partition_item):
-
         node_icon = pyqt.get_icon("mgear_box")
         item_node = self.NODE_CLASS(
             node,
@@ -249,7 +251,6 @@ class PartitionsOutliner(widgets.OutlinerTreeView):
             self._master_item.addChild(child)
 
     def partition_item_enabled_changed(self, item):
-
         if not item:
             return
 
@@ -262,7 +263,6 @@ class PartitionsOutliner(widgets.OutlinerTreeView):
         )
 
     def partition_item_add_skeletal_mesh(self, item):
-
         if not item:
             return
 
@@ -284,7 +284,6 @@ class PartitionsOutliner(widgets.OutlinerTreeView):
         self.reset_contents()
 
     def partition_item_renamed(self, old_name, new_name):
-
         if not old_name or not new_name or old_name == new_name:
             return
 
@@ -295,7 +294,6 @@ class PartitionsOutliner(widgets.OutlinerTreeView):
         export_node.set_partition_name(old_name, new_name)
 
     def partition_skeletal_mesh_removed(self, parent_name, removed_name):
-
         if not parent_name or not removed_name:
             return
 
@@ -312,7 +310,6 @@ class PartitionsOutliner(widgets.OutlinerTreeView):
     def partition_items_dropped(
         self, parent_item, dropped_items, duplicate=False
     ):
-
         if not parent_item or not dropped_items:
             return
